Refactor http_send_buf() and http_recv_header() and simplify buffer
The use of "off" was problematic, because it is state that has to be
handled carefully and could potentially hit edge-cases (e.g. being
larger than sizeof(buf->data)).

The real problem is finding a way to indicate function is "done" with
something (i.e. done filling or sending the buffer). Instead of playing
around with off, which is nothing but a dirty hack, http_send_buf() now
just "drains" the buffer, which also makes it idempotent on an empty
buffer. It indicates to be "done" as soon as the length of the buffer is
zero.

The function http_recv_header() gets an additional paramter "done"
(an int-pointer) which is used to indicate completeness.

As a result, we can drop the off-variable completely from the buffer.

Signed-off-by: Laslo Hunhold <dev@frign.de>