Introduce state-handling and interruptiblity in serve()
Granted, the changes won't make sense right now, but this allows the
functions http_recv_header() and http_send_buf() to be interrupted
(e.g. if a read or write blocks). In the main loop, we currently just
have a do-while-loop, which is not even necessary because the sockets
are blocking and the aforementioned two functions will always complete
their task, unless there's an error.

This change also adapts the code to the new http_recv_header() prototype
with the done-int-pointer.

Signed-off-by: Laslo Hunhold <dev@frign.de>